Back to C main and standalone comdb2ar
