+++ db/comdb2.c
@@ -8560,7 +8560,45 @@ static void getmyid(void)
     gbl_mypid = getpid();
 }
 
-int comdb2_main(int argc, char **argv)
+#define TOOL(x) #x,
+
+#define TOOLS           \
+   TOOL(cdb2_dump)      \
+   TOOL(cdb2_printlog)  \
+   TOOL(cdb2_sqlreplay) \
+   TOOL(cdb2_stat)      \
+   TOOL(cdb2_verify)    \
+   TOOL(cdb2sockpool)   \
+   TOOL(cdb2sql)
+
+#undef TOOL
+#define TOOL(x) int tool_ ##x ##_main(int argc, char *argv[]);
+
+TOOLS
+
+#undef TOOL
+#define TOOL(x) { #x, tool_ ##x ##_main },
+
+struct tool {
+   const char *tool;
+   int (*main_func)(int argc, char *argv[]);
+};
+
+struct tool tool_callbacks[] = {
+   TOOLS
+   NULL
+};
+
+static char* last_path_component(char *argv0) {
+   char *s;
+   s = strrchr(argv0, '/');
+   if (s)
+      return strdup(s+1);
+   else
+      return strdup(argv0);
+}
+
+int main(int argc, char **argv)
 {
     char *marker_file;
     int ii;
@@ -8571,6 +8609,28 @@ int comdb2_main(int argc, char **argv)
     /* clean left over transactions every 5 minutes */
     int clean_mins = 5 * 60 * 1000;
     struct sigaction sact;
+
+    /* allocate initializer first */
+    comdb2ma_init(0, 0);
+
+    /* more reliable */
+#ifdef _LINUX_SOURCE
+    char fname[PATH_MAX];
+    rc = readlink("/proc/self/exe", fname, sizeof(fname));
+    if (rc > 0 && rc < sizeof(fname)) {
+        fname[rc] = 0;
+        exe = last_path_component(fname);
+    }
+#endif
+    if (exe == NULL) {
+       /* more portable */
+       exe = last_path_component(argv[0]);
+    }
+
+    for (int i = 0; tool_callbacks[i].tool; i++) {
+       if (strcmp(tool_callbacks[i].tool, exe) == 0)
+          return tool_callbacks[i].main_func(argc, argv);
+    }
 
     init_debug_switches();
